# api/swagger.yaml
swagger: '2.0'
info:
  description: This is a sample chat server.
  version: 1.0.0
  title: Chat Server Challenge
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: lorenzo@bemind.me
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: chat-server-challenge.herokuapp.com
basePath: /
tags:
  - name: client
    description: Chat Client Challenge
    externalDocs:
      description: Challenge
      url: 'https://github.com/bemindinteractive/chat-client-challenge'
schemes:
  - http
  - https
paths:
  /login:
    post:
      tags:
        - authentication
      summary: Authenticates an user
      description: ''
      operationId: logIn
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Auth object that has to be evaluated
          required: true
          schema:
            $ref: '#/definitions/Auth'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
          headers:
            X-Rate-Limit:
              type: integer
              format: int32
              description: calls per hour allowed by the user
            X-Expires-After:
              type: string
              format: date-time
              description: date in UTC when token expires
        '400':
          description: Invalid username/password supplied
  /logout:
    post:
      tags:
        - authentication
      summary: Logs out current logged in user session
      description: ''
      operationId: logOut
      produces:
        - application/json
      parameters: []
      responses:
        default:
          description: successful operation
  /profile:
    get:
      tags:
        - contacts
      summary: Get currently logged in user data
      description: ''
      operationId: getCurrentUserProfile
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
  /contacts:
    get:
      tags:
        - contacts
      summary: Finds contacts by name
      description: ''
      operationId: findContactsByName
      produces:
        - application/json
      parameters:
        - name: name
          in: query
          description: Status values that need to be considered for filter
          required: false
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '400':
          description: Invalid status value
        '401':
          description: Unauthorized
  '/contacts/{contactId}':
    get:
      tags:
        - contacts
      summary: Get contact by id
      description: ''
      operationId: getContactById
      produces:
        - application/json
      parameters:
        - name: contactId
          in: path
          description: 'The id that needs to be fetched. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid contactId supplied
        '401':
          description: Unauthorized
        '404':
          description: Contact not found
  '/contacts/{contactId}/history':
    get:
      tags:
        - chat
      summary: Get history by contact id
      description: ''
      operationId: getHistoryByContactId
      produces:
        - application/json
      parameters:
        - name: contactId
          in: path
          description: 'The id that needs to be fetched. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/History'
        '400':
          description: Invalid contactId supplied
        '401':
          description: Unauthorized
        '404':
          description: Contact not found
  '/contacts/{contactId}/send':
    post:
      tags:
        - chat
      summary: Get history by contact id
      description: ''
      operationId: sendMessage
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: contactId
          in: path
          description: 'The id that needs to be fetched. '
          required: true
          type: string
        - in: body
          name: body
          description: Message object that has to be sent
          required: true
          schema:
            $ref: '#/definitions/Message'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: Invalid contactId supplied
        '401':
          description: Unauthorized
        '404':
          description: Contact not found
definitions:
  Auth:
    type: object
    required:
      - username
      - password
    properties:
      username:
        type: string
        example: tizio
      password:
        type: string
        example: tizio.secret
  User:
    type: object
    properties:
      id:
        type: string
        example: 2afbf956-42e2-4867-8fb7-2ae06ec096d8
      username:
        type: string
        example: tizio
      name:
        type: string
        example: Tiberio
      surname:
        type: string
        example: Gracco
      email:
        type: string
        example: tizio@bemind.me
      avatar:
        type: string
        example: data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7
  Contact:
    type: object
    properties:
      id:
        type: string
        example: 2afbf956-42e2-4867-8fb7-2ae06ec096d8
      username:
        type: string
        example: tizio
      name:
        type: string
        example: Tiberio
      surname:
        type: string
        example: Gracco
      email:
        type: string
        example: tizio@bemind.me
      avatar:
        type: string
        example: data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7
      history:
        $ref: '#/definitions/History'
  History:
    type: object
    properties:
      messages:
        type: array
        items:
          $ref: '#/definitions/Message'
      unreadCount:
        type: number
        format: int32
        example: 3
  Message:
    type: object
    properties:
      contactId:
        type: string
        example: 76e1b333-757b-4295-9959-6182195d676a
      message:
        type: string
        example: Yo
      sentDate:
        type: string
        format: date
        example: '2017-05-15T13:09:02.035Z'
      readDate:
        type: string
        format: date
        example: '2017-05-15T13:09:02.035Z'